refactor: mission-detail
