+++ src/app/api/missions/[id]/og/route.tsx
@@ -1,6 +1,6 @@
 import { readFile } from "node:fs/promises";
 import { join } from "node:path";
-import { getMissionPageData } from "@/app/missions/[id]/_lib/data";
+import { getMissionPageData } from "@/features/mission-detail/services/mission-detail";
 import { sanitizeImageUrl } from "@/lib/metadata";
 import { Noto_Sans_JP } from "next/font/google";
 import { ImageResponse } from "next/og";
