+++ src/app/missions/[id]/page.tsx
@@ -6,6 +6,9 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
+import { getQuizQuestionsAction } from "@/features/mission-detail/actions/quiz-actions";
+import { MissionWithSubmissionHistory } from "@/features/mission-detail/components/mission-with-submission-history";
+import { getMissionPageData } from "@/features/mission-detail/services/mission-detail";
 import { MissionDetails } from "@/features/missions/components/mission-details";
 import { CurrentUserCardMission } from "@/features/ranking/components/current-user-card-mission";
 import { RankingMission } from "@/features/ranking/components/ranking-mission";
@@ -26,9 +29,6 @@ import { LogIn, Shield } from "lucide-react";
 import type { Metadata } from "next";
 import Link from "next/link";
 import { Suspense } from "react";
-import { MissionWithSubmissionHistory } from "./_components/MissionWithSubmissionHistory";
-import { getMissionPageData } from "./_lib/data";
-import { getQuizQuestionsAction } from "./actions";
 
 type Props = {
   searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
