+++ src/features/mission-detail/components/qr-code-display.tsx
