+++ src/features/mission-detail/components/submission-history.tsx
@@ -1,11 +1,11 @@
 "use client";
 
+import CancelSubmissionDialog from "@/features/mission-detail/components/cancel-submission-dialog";
+import SubmissionItem from "@/features/mission-detail/components/submission-item";
+import { useSubmissionCancel } from "@/features/mission-detail/hooks/use-submission-cancel";
+import type { Submission } from "@/features/mission-detail/types/component-types";
 import type React from "react";
 import { useState } from "react";
-import CancelSubmissionDialog from "./CancelSubmissionDialog";
-import SubmissionItem from "./SubmissionItem";
-import type { Submission } from "./types";
-import { useSubmissionCancel } from "./useSubmissionCancel";
 
 interface SubmissionHistoryProps {
   submissions: Submission[];
