+++ src/features/mission-detail/types/component-types.ts
