+++ src/features/mission-detail/types/detail-types.ts
